pyproject.toml
[project]
name = "iot-anomaly-detection"
version = "0.1.0"
description = ""
authors = [
    {name = "omerrRahsan",email = "farukkoc91@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "duckdb (>=1.2.2,<2.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.8"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
fixable = ["ALL"]


